Add support for Pro FIPS archives
Support Pro archives by a pro property in chisel.yaml. When the pro
property is specified, it has to be either "fips" or "fips-updates". The
repository URL is inferred from the value. Also, credentials for these
URLs are searched. If credentials are not found, the corresponding pro
archive is silently disabled. Otherwise, all requests to the archive are
sent with the credentials found.