Index Release sections by Suite
Currently, the release index (Release file) sections are indexed by the
the Label field value. The label used to find the desired section is the
string "Ubuntu".

That makes it inconvenient to add support for other archives because the
ones we intend to support use different labels, and we'd have to record
them somewhere only for indexing of the release index. Index the release
index by Suite instead, which is already defined for each release index.

Even though the release index is a control file, according to [1], most
indexes we care about contain only a single section. Apt uses the first
section. The control package in chisel does not support sequential
access, and adding it is out of scope for this change, so we have to
index by some field that is present in archives we want to support.
According to [1], none of the defined fields are required, but at least
one must be present. The Suite field was picked for no special reason.

[1] https://wiki.debian.org/DebianRepository/Format#A.22Release.22_files