Make package version selection deterministic
When two or more archives with highest priority contain a package with
the same version, the choice is effectively random, because the
iteration order over options.Archives map is not defined.

Get rid of this nondeterminism by sorting archives by their priority and
then lexicographically by their labels (greater label wins) when the
priorities are equal. Since archive labels are used as keys in YAML map,
they should be unique and the order should be total.