Refactor slicer to lookup by target path
With slice definition like this

  package: pkg
  slices:
    slc:
      /tmp/:
      /tmp/text: {text: foo}

The /tmp/ path is actually included twice in list of /tmp/ target paths,
once as optional, even though both target paths are the same.

To fix this, refactor slicer to maintain package paths in a map keyed by
target path instead of source path.

This will also allow us to maintain relation between target paths and
slices for ChiselDB later. This commit is just a preparation for
following changes ultimately ending with ChiselDB implementation.