Handle implict parent dirs in slicer
Currently we create parent directories in deb/extract.go by testing each
directory path in the tarball for being a prefix of every path to be
extracted from the package. The paths tested include

  a) copy paths,
  b) glob paths and
  c) direct parent directories of paths of other types.

The c) is included by slicer so that with slice definition like this

  package: pkg
  slices:
    sl:
      /tmp/text: {text: foo}
      /var/tmp/text: {text: bar}

the /tmp/, /var/ and /var/tmp/ directories are created from the package
tarball if they exist there. However, only /tmp/ and /var/tmp/ are
included in the paths to extract by slicer. /var/ is created by the
prefix testing in deb/extract.go described above.

This is inconsistent and wasteful. Get rid of the inconsistency by
adding all (non-glob) implicit parent directories in slicer.

Glob path parent directories are not handled (yet), which is not a
problem because they were not handled previously either. They relied and
still rely on os.MkdirAll() in fsutil.Create().