Remove parent directory handling from deb/extract
Before the commit "Handle implict parent dirs in slicer", parent
directories not passed by the slicer in the extract map were created by
iterating every non-glob map value. This code is not used anymore from
slicer.

Remove the code in deb/extract.go that handles parent directories. This
means that unless all parent directories are passed in the extract map,

  - they are created with default attributes (mode 0755) by
    fsutil.Create() or
  - they are not created at all if the path in the extract map is
    optional and doesn't exist in the package.

The first implication now apply to all paths passed in extract map, but
before this commit it already applied to globs. Glob parent directories
were created only in fsutil.Create() by os.MkdirAll() with 0755 mode.

The second implication means that callers will need to pass in all
parent directories they want to create. This is already done in slicer
by the mentioned commit. This commit changes "Optional entries may be
missing" test to account for this new behavior. The old behavior is
preserved in slicer and it's questionable behavior anwyay.