Optimize path testing during extraction
The values in the extract map passed by slicer are looked up by linear
search over map entries. Further, each path in the extract map is tested
for * and ? on every path encountered in the package.

Filter the globs into array before we scan the package, sort the array
for deterministic behavior, and lookup paths in the extract map directly
with fall back to linear search of globs only.