Bring back parent directory handling to deb/extract
This commit makes deb/extract create parent directories again. This time
even for globs!

Since the order of the tar scan is not defined, when target path parents
were not encountered before, they are created with 0755 and later
chmoded to the proper mode when they are found in the tarball.
Otherwise, they stay 0755. This may happen when the target path is in a
directory that's not present in the package.

The new code does not rely on os.MkdirAll() calls in fsutil.Create() (it
works without it), but there are other call sites which expects parents
to be created in fsutil.Create(), namely, archive tests.