fix: add ``--timeout`` option in cut command
Add ``--timeout`` option in chisel cut so that users can
specify their preferred limit for each request to avoid
context deadline error [0].

Duration (with unit) should be specified as the argument of
the new option. A timeout of zero means no timeout. Default
timeout is 60 seconds.

Additionally, increase the hardcoded Timeout in the http
client in archive.go to match the default 60 seconds limit.

Fixes #14.

References:
- [0] https://github.com/canonical/chisel/issues/14