fix(cmd/find): change query matching logic
This commit changes the logic of matching a query against a slice.
Following the new behaviour, a slice name will be listed as a match if
the query is a sub-string of the slice name, in pkg_slice format.
Or, it will be listed if the query is atmost 1 (One) Levenshtein
distance [1] away from the slice name, in pkg_slice format.

[1] https://en.wikipedia.org/wiki/Levenshtein_distance