refactor: move release reading logic to a function
Previously, release info (setup.Release) parsing/reading logic was
written in cmd/chisel/cmd_cut.go file. Being the only operation command
before, it made sense to have it there. But, since there are incoming
new commands, it is likely that this logic will be re-used. Thus, move
it to a function, ``getRelease()``.