feat: add info command
This commit adds a new ``info`` command which shows detailed information
about package slices.

It accepts a white-space separated list of strings. The list can be
composed of package names, slice names, or a combination of both. The
default output format is YAML. When multiple package or slice names are
provided, the output is a list of YAML documents, separated by three
dashes (“---”).

Slice definitions are shown verbatim according to their definition in
the Chisel releases. For example, "globs" are not expanded.

---

Usage:
  chisel info [info-OPTIONS] [<pkg|slice>...]

[info command options]
      --release=<branch|dir>      Chisel release branch or directory