Verify release signatures
This commit builds on the previous commit that introduces support for
OpenPGP keyrings to verify signatures of archive release files.
Previously, we used the Release file of each configured suite. With this
commit, we fetch the InRelease file instead. This file is in clearsign
format[1] and contains the Release file and its signature[2].

[1] https://www.gnupg.org/gph/en/manual/x135.html
[2] https://www.chiark.greenend.org.uk/~cjwatson/blog/no-more-hash-sum-mismatch-errors.html