feat: integrity checks for Ubuntu Release files (#106)
This adds the functionalities to add public keys to validate archive InRelease file signatures. Instead of downloading the "Release" files for different suites, download the signed "InRelease" file and validate the signature with the provided key(s).

This also extends the chisel release with public keys definitions. They are stored in ASCII armored format in the top-level public-keys property by name (v1-public-keys during the transition). They are then referenced by name in the public-keys list property in archive definitions.