bugfix: parent directory permissions
This commit fixes three bugs regarding parent directory creation.

The first one is that implicit folders were overwriting the permissions
of explicit ones in slice definitions (e.g. slice A declares /foo/bar
and slice B /foo, if slice A was installed after B then the implicit
permissions of /foo from A would prevail.

The second one is that we were creating parent directories of optional
files which did not exist.

The third bug is that we were not using the permissions from the tarball
for parent directories if the path was a glob.