move cmd_cut logic to new cut package
The slicer package will only be concerned about installing a selection
of slices which will allow us to use it and test it in isolation. The
main benefit is that it decouples slice installation and manifest
creation and archive fetching, which we might want to do differently for
future use-cases such as chiselling an already chiselled fs.

There is a new cut package that contain all the logic for cutting.