refactor: filter created entries for report  (#124)
To make the content transparent and facilitate changing the package,
this commit creates the deb used for testing programmatically instead of
embedding it as a base64-encoded blob directly.

Also, the entries not mentioned explicitly in the slice definition will not be
added to the report (e.g. copyright, parent directories). Additionally,
we have removed the Globbed option and simplified it.