bugfix: multiple slices of same package produces faulty report
In slicer.go, the slice paths are grouped by package names and then for
each package the overall contents are extracted together. While this is
fine, due to existing code, it would appear as in all of those paths are
coming from a single slice. Later, when adding the paths to the report,
it fails the check that those paths are truly not coming from that
slice. Hence, those paths do not get added to the report.

For example, in the following slice definition, if the "manifest" slice
appears first in the sorted order, the path "/dir/file" does not get
added to the report. Although it is extracted, it is not coming from
"manifest" -- thus the check fails and it is not added to report later.

    package: test-package
    slices:
	myslice:
	    contents:
		/dir/file:
		/dir/file-copy:  {copy: /dir/file}
		/other-dir/file: {symlink: ../dir/file}
		/dir/text-file:  {text: data1}
		/dir/foo/bar/:   {make: true, mode: 01777}
	manifest:
	    contents:
		/var/lib/foo/**: {generate: manifest}

This commit fixes the bug.