feat: add mutated files' properties
The paths with ``mutable: true`` property can be "mutated" with mutation
scripts. Thus, the files matching those paths may have a different size
and hash after the mutation scripts have been run. The report and DB
should reflect these changes by updating the entries of the mutated
paths.

Paths can also have the ``until: mutate`` property. This means that
those files are available until the mutation scripts have been run and
are not present in the final file system. Thus, they should not be part
of the report or the DB either.

This commit adds support for both -- it updates the changed properties
of mutated files and makes sure not to add report entries for
``until: mutate``.

Co-authored-by: Alberto Carretero <angelalbertoc.r@gmail.com>