feat: support multiple archives with "priority" field (#160)
This commit adds support for fetching packages from multiple archives. It introduces a new field archives.<archive-name>.priority which takes in a signed integer and specifies the priority of a certain archive. A package is fetched from the archive with highest priority and negative priorities are ignored unless explicitly pinned in a package.

The concept of default archive in chisel.yaml is now deprecated. However, the field can still be present and it will be parsed and interpreted in a compatible way for the time being.

DEPRECATED: "archives.*.default" field in chisel.yaml.