feat: debug command to check release archives
This commit introduces the `chisel debug check-release-archives` command
that verifies that Chisel assumptions avoid the archives hold. For
example, Chisel assumes that parent directories bundled in several
packages in the archive are consistent with one another (i.e. have the
same mode). If the is not the case, we need to flag the issue and
possibly handle it in the slice definitions files.