feat: support unmaintained and unstable Ubuntu releases (#238)
This commits adds maintenance status to a release as the top-level field maintenance in chisel.yaml. The dates there specify the lifecycle of the release in regards to its support status, with and without Ubuntu Pro subscriptions. This enables Chisel to fail if the release and its combination of archives is no longer supported officially and it may lead to security problems. For example, interim releases like mantic and no longer officially supported, and LTSs like focal are only maintained when using ESM which needs an Ubuntu Pro subscription. If no maintained archive is available Chisel will fail, but the user can continue execution by passing --ignore=unmaintained as a CLI flag.

When the dates are into the future it indicates instead that the release is unstable. The user can choose to continue the execution by passing --ignore=unstable as a CLI flag.

Lastly, this commits also allows Chisel to work with interim releases by using the dates described above to change the archive's URL from archive.ubuntu.com to old-releases.ubuntu.com when needed.

Compatibility-wise, Chisel has a list of releases with their default dates so it continues to work with existing chisel.yaml. There will be a window when maintenance is optional for these releases and chisel-releases will be updated. Then, the field will be mandatory.

IMPORTANT: This new version of Chisel will fail to cut Ubuntu 20.04 when Ubuntu Pro is not enabled because the standard archive is unmaintained. Users need to add --ignore=unmaintained or move to a more recent release or use the Ubuntu Pro archives.