bugfix: revert change to transitive dependencies (essential)
This commit reverts the change in
aca3b32315cda2b5e4c2e9854bf249167e328385 that changed the for loop to
use the new range semantics which was not equivalent because we were
mutating the contents of the array in the loop itself.

A test has been added to detect it.